Validated the fields of Beer and checked the saving and updating methods with postman

+++ spring-rest-webservice/src/main/java/com/udemy/springframework/repositories/BeerRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 
 import com.udemy.springframework.entities.Beer;
 
-public interface BeerRepository extends JpaRepository<Beer, Long> {
+public interface BeerRepository extends JpaRepository<Beer, Integer> {
 }
